서비스 계층 단위 테스트 작성 완료

+++ src/main/java/Nengcipe/NengcipeBackend/service/MemberRecipeService.java
@@ -5,6 +5,7 @@ import Nengcipe.NengcipeBackend.domain.MemberRecipe;
 import Nengcipe.NengcipeBackend.domain.Recipe;
 import Nengcipe.NengcipeBackend.dto.MemberRecipeRequestDto;
 import Nengcipe.NengcipeBackend.dto.ResultResponse;
+import Nengcipe.NengcipeBackend.exception.DuplicationException;
 import Nengcipe.NengcipeBackend.exception.NotFoundException;
 import Nengcipe.NengcipeBackend.repository.MemberRecipeRepository;
 import Nengcipe.NengcipeBackend.repository.MemberRepository;
@@ -33,12 +34,15 @@ public class MemberRecipeService {
      * @return newMemberRecipe
      */
     public MemberRecipe createScrapRecipe(Member member, Recipe recipe){
-        MemberRecipe newMemberRecipe = MemberRecipe.myMemberRecipe()
+        Optional<MemberRecipe> delMemberRecipe = memberRecipeRepository.findByMemberAndRecipe(member, recipe);
+        if (delMemberRecipe.isPresent()) {
+            throw new DuplicationException("스크랩 레시피", null);
+        }
+        MemberRecipe newMemberRecipe = MemberRecipe.builder()
                 .member(member) //member_id에 해당하는 member 객체 (jwt)
                 .recipe(recipe) //recipe_id에 해당하는 recipe 객체 (findbyid)
                 .build();
-        memberRecipeRepository.save(newMemberRecipe); //newMemberRecipe란 객체를 저장소에 저장하는 역할
-        return newMemberRecipe;
+        return memberRecipeRepository.save(newMemberRecipe); //newMemberRecipe란 객체를 저장소에 저장하는 역할
 
     }
 
@@ -52,6 +56,9 @@ public class MemberRecipeService {
      */
     public MemberRecipe deleteScrapRecipe(Member member, Recipe recipe) {
         Optional<MemberRecipe> delMemberRecipe = memberRecipeRepository.findByMemberAndRecipe(member, recipe);
+        if (delMemberRecipe.isEmpty()) {
+            throw new NotFoundException("스크랩 레시피", null);
+        }
         memberRecipeRepository.delete(delMemberRecipe.get());
         return delMemberRecipe.get();
     }
